## May 7, 2019
* TOC
{: toc }

### Essay ideas from early April

I sometimes go through phases of creativity where I start a bunch of essays or talks that I never get around to finishing. This happened early this past August. But I had *so* many ideas that I didn't even have time to finish *outlining* them all. I took an hour today to get those outlines in here *without cleaning them up much so they will mostly be nonsense to anyone but me*. I created separate files for these outlines:

* [Tentative derivation of compositionality as the primary subgoal of PL/HCI design](/drafts/composability)
* [Natural Langauge is not and will never be the future of coding](/drafts/natural-langauge)

The rest of them were just small stub outlines so I'll put them right here in the log:

#### Where is the line between programming vs using a computer?  (stub outline)
  
<blockquote class="twitter-tweet" data-lang="en"><p lang="en" dir="ltr">Programming is setting up computation over time. Unfortunately this means that many important insights about your program aren&#39;t apparent until future times  Live programming is about bringing insights from future times to the time of programming</p>&mdash; Steve Krouse (@stevekrouse) <a href="https://twitter.com/stevekrouse/status/1113813390886100994?ref_src=twsrc%5Etfw">April 4, 2019</a></blockquote>

And this thread: https://twitter.com/stevekrouse/status/1084424882765520897

#### What is programming?  (stub outline)
  
<blockquote class="twitter-tweet" data-conversation="none" data-lang="en"><p lang="en" dir="ltr">For what it&#39;s worth, I really like the &quot;soft&quot; part. My vision for software is that it &quot;feels like clay&quot;, is infinitely customizable, the sky is the limit. It is the stuff of thoughts, not of atoms. Fiction, not nonfiction</p>&mdash; Steve Krouse (@stevekrouse) <a href="https://twitter.com/stevekrouse/status/1114898545910456321?ref_src=twsrc%5Etfw">April 7, 2019</a></blockquote>

Programming is the study of precision like math is the study of long chains of close reasoning without paradoxes and history is the study of past humans through text.  
  
#### Learning to Code Isn't a Thing  (stub outline)
  
* just like learning to speak a language (or use a computer) isn't a thing  
* personalization and intrinsic motivation is key, then env feedback loop, easy setup, googleability, unstuckability
* 
<blockquote class="twitter-tweet" data-lang="en"><p lang="en" dir="ltr">When helping someone learn to code, I ask them what they want to be able to code. Then I help them find an environment (coding notebook, online IDE, desktop IDE, Zapier, spreadsheet, etc) with the best feedback loops, easiest setup. The language is not part of the equation <a href="https://t.co/aNwGdnIkmd">https://t.co/aNwGdnIkmd</a></p>&mdash; Steve Krouse (@stevekrouse) <a href="https://twitter.com/stevekrouse/status/1114649196756385793?ref_src=twsrc%5Etfw">April 6, 2019</a></blockquote>
  
#### Immutable Code Editing (stub outline)
  
editor = f(node, editState) so f can be a recursive polymorphic func... cyrus and hazel... cursor is key<  
  
Recursiveness of the state = lift validatorfunction allModifyerFunctionstreams initial state  
  
But then of course each piece of this defition is a behavior with it's own validator function that's a behavior with it's own validator function...  
  
Ultimately this Programming Language is one single fucking immutable expression! It is Behavior [Expressions] (aside: I want Behavior length list where each item is Behavior) defined as (fold processEdit (fold concat newExpression []) Behavior [self]  

[This SVG is hard to read but shows a bit of the thinking here.](https://gistcdn.githack.com/stevekrouse/7c1d28c1106dedcfd72a992c3e471460/raw/8172c6df7afdd4208d6d2819c052c013329ae7d3/immutable%20code%20editing.svg)
  
#### How many language features can be replaced with editor ones? (stub outline)

* lets, closures, imports (Cyrus Omar of Hazel talks about this)
* it seems like most things can be editor features, especially if the editor is recognizing patterns and making syntactic sugar out of them (a la lamdu)
* it's the things you want in the AST (stored as metadata) that maybe should be in the language semantics?

### Week of May 5th, 2019 (and beyond)

Yesterday I cleaned up my email, todo lists, random errands, and went on a run. I also had dinner with Sam John of Hopscotch, which was fun.

Today I did more errands, a bit of consulting work for TCS, and adding things to the website I've been meaning to for a while. I also cleaned up my workflowy which feels great.

My meeting with Dark was pushed to next week, which frees me up to work on my research proposal draft 2 this week, which is due on Friday and then I'll speak with JE afterwards about next steps.

Next week the focus will be Zapier for Dark as well as the draft of the piece framing the research we will be publishing on my site. So it'll be Dark-heavy next week. I'll also be traveling back to London Weds, so won't get much work done then or later in the week as I readjust, so let's just say Dark and random errands is all next week.

The following week I might try and (finally) publish the Cyrus Omar episode and edit a few others. I'll also want to get back to research, as well as doing the work for TCS on curriculum research I promised.

So many balls in the air! It's key to keep my organizational systems clean to manage it all. I really feel like I need the malleable software medium to really build what I need here to make my brain powerful enough to handle it.