## Friday Sept 27, 2019
* TOC
{: toc }

### JE Call on 9/25/19

On Wednesday I spent two hours preparing for the call with JE, thinking through my plan for a "vision statement" project. The notes from the call are [here](/notes/jonathan-edwards/09-25-19). He approved of the plan and suggested this year's Salon de Refuge as a place to publish these notes. We agreed on **next Friday, Oct 4** as the time I will send him a couple paragraphs of a plan for this project:

- my research goals, what I'm doing and why
- write the abstract to the paper I want to submit

Here are the prep notes:

- Are there similar efforts like this? Clearly the Mother of All Demos comes to mind. But the Internet, Wikipedia, and open-source were all existence proofs. Yes some people hyped them up but it just took time for them to proof in the pudding. BV's talks come to mind but they also have a mixed legacy. They got me (and Webflow and Notion and Explorables) so I think its net super positive.

- Markers of success	
	- virality: super top of HN/reddit/twitter with thousands of upvotes, retweets
	- people citing it as influential to their own work (or career changes)
	- friends and family better understand what I'm trying to do (more incidental)
	- people want to sponsor or invest in what I'm doing (more inbound support requests)

- Goal take away messages
	- something about how there are all sorts of limitations with technology that we have learned to live with but that are worth trying to fix
	- a feeling that the computer revolution that has happened doesn't live up to what can happen
	- a sense of medium from speaking to writing to printing press, radio, tv, internet, social media, etc that is so key and merits focus
	- wow, these are similar to https://futureofcoding.org/drafts/statement#5-year-goal!
		- in other words, my new thinking is: can I demo this system without a building it? can I use my imagination to cut corners?

- Day in the life
	 - The problem with demo-ing using such a system without modifying it is that it feels like another app. The problem with modifying it too much is that it feels like a lot of work, when in reality the modifications happen rarely
	- Some interfaces I wish existed
		- I could train my phone or watch with custom voice commands, such as: "I will follow up on that" to retroactively save and remind me about the last couple minutes of conversation.  Trying to solve the problem of having to pull out my phone to take notes or reminders and disconnect from the conversation.
		- The ability to connect to multiple wifi networks at the same time, or dynamically switch between them based on bandwidth
		- Block phone numbers based on a pattern (app for this)
		- Spotify running
		- mix and match bits of functionality from google apps and code
		- todo list, email, amazing marvin, team project management
		- CRMs, Notion, Coda
		- invite list for party, expected attendance, tracking responses, seating list, analytics
		- https://futureofcoding.org/drafts/ulf-2019
		- expense tracking

- Other idea is to do interviews with people to get them to imagine / pretend build in Figma what interfaces they'd want

Let's get visual! This system exists. What do I build?

The system would let me interact with my Inbox as a stream of emails of some sort. I think that the idea of an "inbox" that you try and keep to zero by either doing the thing or putting it in a bucket of whent to do it is a good strategy. All emails (accept spam or others filtered out) will go to this inbox. I will also put new todos in here that are generated in real life such as meetings. This is like emailing yourself todos. --> One way to think of this strategy is emailing yourself todos and snoozing things. That works but it doesn't allow reordring of todos or nesting of todos or a lot of other expression. Something closer to Coda is what I want but also email within it. But you can do that with Coda! You can even view the emails one at a time in something that looks like Gmail for iPad. But it's not nearly as nice... This is where it gets subtle. I want to be able to have everything how I want it but I don't want to have to make it all myself. I'd like to customize building on what others have done, occasionally adding my own little customizations, mostly through composability, and occasionally through work from scratch. I think I have to bite the bullet and actually design it... Here's a ~20 minutes of work on some sketches:

![](https://user-images.githubusercontent.com/2288939/65612444-2df65c80-dfac-11e9-9f44-029f3d4bcf6b.png)

Places I drew from and shouldn't forget when doing this work:

- https://futureofcoding.org/log#end-user--end-community-computing-vision
- https://futureofcoding.org/drafts/statement#5-year-goal
- https://futureofcoding.org/drafts/ulf-2019

### Other vision statement thoughts

- How we allow for the functionality of Uber (people to exchange money for rides) without the centralized company Uber? For example, drivers can export their location, verified reviews, and a pricing function. Can we generalize for all sorts of such exchanges in a Craigslist sort of way? How do people export the _type_ of service they provide? Do we need an ontology of services and products or can simple searching work? 
- Metaphorically, Illich's thoughts on bikes vs cars could be an interesting way to explain end-user programming vs corporate-built software. It's similar to the metaphor I got from Dynamicland of a knife vs a Nespresso, which recently worked well on a friend.